requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
rembg[cpu]==2.0.62
pillow==11.1.0
python-multipart==0.0.18
numpy>=1.24.0
onnxruntime>=1.17.0
